+++ src/components/badge/badge.styles.ts
@@ -26,27 +26,27 @@ export default css`
   /* Type modifiers */
   .badge--primary {
     background-color: rgb(var(--sl-color-primary-500));
-    color: rgb(var(--sl-color-primary-text));
+    color: rgb(var(--sl-color-white));
   }
 
   .badge--success {
     background-color: rgb(var(--sl-color-success-500));
-    color: rgb(var(--sl-color-success-text));
+    color: rgb(var(--sl-color-white));
   }
 
   .badge--info {
     background-color: rgb(var(--sl-color-info-500));
-    color: rgb(var(--sl-color-info-text));
+    color: rgb(var(--sl-color-white));
   }
 
   .badge--warning {
     background-color: rgb(var(--sl-color-warning-500));
-    color: rgb(var(--sl-color-warning-text));
+    color: rgb(var(--sl-color-white));
   }
 
   .badge--danger {
     background-color: rgb(var(--sl-color-danger-500));
-    color: rgb(var(--sl-color-danger-text));
+    color: rgb(var(--sl-color-white));
   }
 
   /* Pill modifier */